Refactor and add jPro
This refactors the project into a package. It also adds the jPro dependency. Because
of this Java was changed from 8 to 11. JavaFX is a separate SDK with 11 so that has
been included in the project. You must add it as a library to your project. You must
also pass --module-path "path\to\javafx-sdk-11.0.2\lib" --add-modules javafx.controls,javafx.fxml
to the VM options in the run configurations.